requirements.txt
Flask>=0.10.1
lxc-python2==0.1
